move conversion to connector
