+++ src/common/Cargo.toml
@@ -59,10 +59,6 @@ itoa = "1.0"
 jsonbb = { workspace = true }
 lru = { workspace = true }
 memcomparable = { version = "0.2", features = ["decimal"] }
-mysql_async = { workspace = true }
-mysql_common = { version = "0.32", default-features = false, features = [
-    "chrono",
-] }
 num-integer = "0.1"
 num-traits = "0.2"
 number_prefix = "0.4.0"
