+++ src/common/src/lib.rs
@@ -80,7 +80,6 @@ pub use {
     risingwave_license as license,
 };
 pub mod lru;
-pub mod mysql;
 pub mod opts;
 pub mod range;
 pub mod row;
